+++ docs/resources/changelog.md
@@ -6,6 +6,10 @@ Components with the <sl-badge variant="warning" pill>Experimental</sl-badge> bad
 
 _During the beta period, these restrictions may be relaxed in the event of a mission-critical bug._ 🐛
 
+## Next
+
+- Revert menu item caching due to regression [#766](https://github.com/shoelace-style/shoelace/issues/766)
+
 ## 2.0.0-beta.74
 
 - 🚨 BREAKING: reworked focus rings to use outlines instead of box shadows